+++ src/main/java/com/telepathicgrunt/the_bumblezone/configs/BzDungeonsConfigs.java
@@ -15,23 +15,23 @@ public class BzDungeonsConfigs {
             builder.push("Dungeon Options");
 
                 beeDungeonRarity = subscriber.subscribe(builder
-                    .comment(" \r\n-----------------------------------------------------\r\n\r\n"
-                            +" How rare Bee Dungeons are. Higher numbers means more rare.\r\n"
-                            +" Default rate is 1. Setting to 1001 will disable Bee Dungeons.\r\n")
+                    .comment(" \n-----------------------------------------------------\n",
+                           " How rare Bee Dungeons are. Higher numbers means more rare.",
+                           " Default rate is 1. Setting to 1001 will disable Bee Dungeons.\n")
                     .translation("the_bumblezone.config.dungeons.beedungeonrarity")
                     .defineInRange("beeDungeonRarity", 1, 1, 1001));
 
                 spiderInfestedBeeDungeonRarity = subscriber.subscribe(builder
-                    .comment(" \r\n-----------------------------------------------------\r\n\r\n"
-                            +" How rare Spider Infested Bee Dungeons are. Higher numbers means more rare.\r\n"
-                            +" Default rate is 8. Setting to 1001 will disable Bee Dungeons.\r\n")
+                    .comment(" \n-----------------------------------------------------\n",
+                           " How rare Spider Infested Bee Dungeons are. Higher numbers means more rare.",
+                           " Default rate is 8. Setting to 1001 will disable Bee Dungeons.\n")
                     .translation("the_bumblezone.config.dungeons.spiderinfestedbeedungeonrarity")
                     .defineInRange("spiderInfestedBeeDungeonRarity", 8, 1, 1001));
 
                 spawnerRateSpiderBeeDungeon = subscriber.subscribe(builder
-                    .comment(" \r\n-----------------------------------------------------\r\n\r\n"
-                            +" How rare are Spider/Cave Spider Spawners in Spider Infested Bee Dungeons.\r\n"
-                            +" 0 is no spawners, 1 is maximum spawners, and default is 0.2D\r\n")
+                    .comment(" \n-----------------------------------------------------\n",
+                           " How rare are Spider/Cave Spider Spawners in Spider Infested Bee Dungeons.",
+                           " 0 is no spawners, 1 is maximum spawners, and default is 0.2D\n")
                     .translation("the_bumblezone.config.dungeons.spawnerratespiderbeedungeon")
                     .defineInRange("spawnerRateSpiderBeeDungeon", 0.2D, 0D, 1D));