+++ src/main/java/com/telepathicgrunt/the_bumblezone/configs/BzModCompatibilityConfigs.java
@@ -58,14 +58,14 @@ public class BzModCompatibilityConfigs
 					builder.push("Caves and Cliffs Backport Options");
 
 					allowCACCandlesBeeDungeon = subscriber.subscribe(builder
-							.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-									+" Spawn Caves and Cliffs Backport Candles in Bee Dungeons.\r\n")
+							.comment(" \n-----------------------------------------------------\n\n"
+									+" Spawn Caves and Cliffs Backport Candles in Bee Dungeons.\n")
 							.translation("the_bumblezone.config.modcompat.cavesandcliffs.allowcaccandlesbeedungeon")
 							.define("allowCACCandlesBeeDungeon", true));
 
 					allowCACCandlesSpiderBeeDungeon = subscriber.subscribe(builder
-							.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-									+" Spawn Caves and Cliffs Backport Candles in Spider Infested Bee Dungeons.\r\n")
+							.comment(" \n-----------------------------------------------------\n\n"
+									+" Spawn Caves and Cliffs Backport Candles in Spider Infested Bee Dungeons.\n")
 							.translation("the_bumblezone.config.modcompat.cavesandcliffs.allowcaccandlesspiderbeedungeon")
 							.define("allowCACCandlesSpiderBeeDungeon", true));
 
@@ -74,20 +74,20 @@ public class BzModCompatibilityConfigs
 					builder.push("Charm Options");
 
 					allowCCandlesBeeDungeon = subscriber.subscribe(builder
-							.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-									+" Spawn Charm Candles in Bee Dungeons.\r\n")
+							.comment(" \n-----------------------------------------------------\n\n"
+									+" Spawn Charm Candles in Bee Dungeons.\n")
 							.translation("the_bumblezone.config.modcompat.charm.allowccandlesbeedungeon")
 							.define("allowCCandlesBeeDungeon", true));
 
 					allowCCandlesSpiderBeeDungeon = subscriber.subscribe(builder
-							.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-									+" Spawn Charm Candles in Spider Infested Bee Dungeons.\r\n")
+							.comment(" \n-----------------------------------------------------\n\n"
+									+" Spawn Charm Candles in Spider Infested Bee Dungeons.\n")
 							.translation("the_bumblezone.config.modcompat.charm.allowccandlesspiderbeedungeon")
 							.define("allowCCandlesSpiderBeeDungeon", true));
 
 					allowCharmTradeCompat = subscriber.subscribe(builder
-							.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-									+" Add new trades to Charm's Beekeeper Villager.\r\n")
+							.comment(" \n-----------------------------------------------------\n\n"
+									+" Add new trades to Charm's Beekeeper Villager.\n")
 							.translation("the_bumblezone.config.modcompat.charm.allowcharmtradecompat")
 							.define("allowCharmTradeCompat", true));
 
@@ -96,8 +96,8 @@ public class BzModCompatibilityConfigs
 					builder.push("Pokecube Options");
 
 					spawnPokecubeBeePokemon = subscriber.subscribe(builder
-							.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-									+" Spawn Pokecube's bee-like pokemon in The Bumblezone and from Honey Brood Blocks.\r\n")
+							.comment(" \n-----------------------------------------------------\n\n"
+									+" Spawn Pokecube's bee-like pokemon in The Bumblezone and from Honey Brood Blocks.\n")
 							.translation("the_bumblezone.config.modcompat.pokecube.spawnpokecubebeepokemon")
 							.define("spawnPokecubeBeePokemon", true));
 
@@ -106,81 +106,81 @@ public class BzModCompatibilityConfigs
 					builder.push("Resourceful Bees Options");
 
 					spawnResourcefulBeesBeesMob = subscriber.subscribe(builder
-							.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-									+" Spawn Resourceful Bees in The Bumblezone and from Honey Brood Blocks\r\n"
-									+" alongside regular bees at a 1/15th chance when spawning regular bees.\r\n")
+							.comment(" \n-----------------------------------------------------\n\n"
+									+" Spawn Resourceful Bees in The Bumblezone and from Honey Brood Blocks\n"
+									+" alongside regular bees at a 1/15th chance when spawning regular bees.\n")
 							.translation("the_bumblezone.config.modcompat.resourcefulbees.spawnresourcefulbeesbeesmob")
 							.define("spawnResourcefulBeesBeesMob", true));
 
 					spawnResourcefulBeesHoneycombVariants = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" Spawn Resourceful Bees's various honeycomb variants in The Bumblezone\r\n"
-								+" at all kinds of heights and height bands. Start exploring to find \r\n"
-								+" where they spawn!\r\n"
-								+" \r\n"
-								+" NOTE: Will require a restart of the world to take effect. \r\n")
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" Spawn Resourceful Bees's various honeycomb variants in The Bumblezone\n"
+								+" at all kinds of heights and height bands. Start exploring to find \n"
+								+" where they spawn!\n"
+								+" \n"
+								+" NOTE: Will require a restart of the world to take effect. \n")
 						.translation("the_bumblezone.config.modcompat.resourcefulbees.spawnproductivebeeshoneycombvariants")
 						.define("spawnResourcefulBeesHoneycombVariants", true));
 
 					RBBeesWaxWorldgen = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" Spawn Resourceful Bees's Wax Block as part of The Bumblezone's worldgen.\r\n")
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" Spawn Resourceful Bees's Wax Block as part of The Bumblezone's worldgen.\n")
 						.translation("the_bumblezone.config.modcompat.resourcefulbees.rbbeeswaxworldgen")
 						.define("RBBeesWaxWorldgen", true));
 
 					useSpawnInWorldConfigFromRB = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" Use Resourceful Bees's canSpawnInWorld config on their bee data\r\n"
-								+" to know what bees to spawn in Bumblezone. This will stack with\r\n"
-								+" RBBlacklistedBees config entry that Bumblezone uses.\r\n"
-								+" Bees blacklisted from either will not spawn and their combs will not spawn either.\r\n"
-								+" \r\n"
-								+" NOTE: Will require a restart of the world to take effect. \r\n")
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" Use Resourceful Bees's canSpawnInWorld config on their bee data\n"
+								+" to know what bees to spawn in Bumblezone. This will stack with\n"
+								+" RBBlacklistedBees config entry that Bumblezone uses.\n"
+								+" Bees blacklisted from either will not spawn and their combs will not spawn either.\n"
+								+" \n"
+								+" NOTE: Will require a restart of the world to take effect. \n")
 						.translation("the_bumblezone.config.modcompat.resourcefulbees.usespawninworldconfigfromrb")
 						.define("useSpawnInWorldConfigFromRB", false));
 
 					RBBlacklistedBees = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" Blacklist what Resourceful Bees bees should not spawn in Bumblezone. \r\n"
-								+" Separate each entry with a comma. Example: \"resourcefulbees:iron,resourcefulbees:coal\"\r\n"
-								+" \r\n"
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" Blacklist what Resourceful Bees bees should not spawn in Bumblezone. \n"
+								+" Separate each entry with a comma. Example: \"resourcefulbees:iron,resourcefulbees:coal\"\n"
+								+" \n"
 								+" Note: Blacklisted bees will automatically blacklist their respective combs from worldgen too.")
 						.translation("the_bumblezone.config.resourcefulbees.rbblacklistedbees")
 						.define("RBBlacklistedBees", ""));
 
 					RBOreHoneycombSpawnRateBeeDungeon = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" How much of Bee Dungeons is made of ore-based honeycombs.\r\n"
-								+" 0 is no or honeycombs, 1 is max ore honeycombs, and default is 0.3D\r\n")
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" How much of Bee Dungeons is made of ore-based honeycombs.\n"
+								+" 0 is no or honeycombs, 1 is max ore honeycombs, and default is 0.3D\n")
 						.translation("the_bumblezone.config.resourcefulbees.rborehoneycombspawnratebeedungeon")
 						.defineInRange("RBOreHoneycombSpawnRateBeeDungeon", 0.3D, 0D, 1D));
 
 					RBGreatHoneycombRarityBeeDungeon = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" How rare good ore-based Honeycombs (diamonds, ender, emerald, etc) are \r\n"
-								+" in Bee Dungeons. \r\n"
-								+" Higher numbers means more rare. Default rate is 3.\r\n")
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" How rare good ore-based Honeycombs (diamonds, ender, emerald, etc) are \n"
+								+" in Bee Dungeons. \n"
+								+" Higher numbers means more rare. Default rate is 3.\n")
 						.translation("the_bumblezone.config.resourcefulbees.rbgreathoneycombraritybeedungeon")
 						.defineInRange("RBGreatHoneycombRarityBeeDungeon", 2, 1, 1001));
 
 					RBOreHoneycombSpawnRateSpiderBeeDungeon = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" How much of Spider Infested Bee Dungeons is made of ore-based honeycombs.\r\n"
-								+" 0 is no or honeycombs, 1 is max ore honeycombs, and default is 0.1D\r\n")
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" How much of Spider Infested Bee Dungeons is made of ore-based honeycombs.\n"
+								+" 0 is no or honeycombs, 1 is max ore honeycombs, and default is 0.1D\n")
 						.translation("the_bumblezone.config.resourcefulbees.rborehoneycombspawnratespiderbeedungeon")
 						.defineInRange("RBOreHoneycombSpawnRateSpiderBeeDungeon", 0.1D, 0D, 1D));
 
 					RBGreatHoneycombRaritySpiderBeeDungeon = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" How rare good ore-based Honeycombs (diamonds, ender, emerald, etc) are \r\n"
-								+" in Spider Infested Bee Dungeons. \r\n"
-								+" Higher numbers means more rare. Default rate is 2.\r\n")
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" How rare good ore-based Honeycombs (diamonds, ender, emerald, etc) are \n"
+								+" in Spider Infested Bee Dungeons. \n"
+								+" Higher numbers means more rare. Default rate is 2.\n")
 						.translation("the_bumblezone.config.resourcefulbees.rbgreathoneycombrarityspiderbeedungeon")
 						.defineInRange("RBGreatHoneycombRaritySpiderBeeDungeon", 2, 1, 1001));
 
 					allowResorucefulBeesTradeCompat = subscriber.subscribe(builder
-							.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-									+" Add new trades to Resourceful Bees's Beekeeper Villager.\r\n")
+							.comment(" \n-----------------------------------------------------\n\n"
+									+" Add new trades to Resourceful Bees's Beekeeper Villager.\n")
 							.translation("the_bumblezone.config.modcompat.resourcefulbees.allowresorucefulbeestradecompat")
 							.define("allowResorucefulBeesTradeCompat", true));
 
@@ -189,68 +189,68 @@ public class BzModCompatibilityConfigs
 				builder.push("Productive Bees Options");
 
 					spawnProductiveBeesBeesMob = subscriber.subscribe(builder
-		                    .comment(" \r\n-----------------------------------------------------\r\n\r\n"
-		                    		+" Spawn Productive Bees in The Bumblezone and from Honey Brood Blocks\r\n"
-		                    		+" alongside regular bees at a 1/15th chance when spawning regular bees.\r\n")
+		                    .comment(" \n-----------------------------------------------------\n\n"
+		                    		+" Spawn Productive Bees in The Bumblezone and from Honey Brood Blocks\n"
+		                    		+" alongside regular bees at a 1/15th chance when spawning regular bees.\n")
 		                    .translation("the_bumblezone.config.modcompat.productivebees.spawnproductivebeesbeesmob")
 		                    .define("spawnProductiveBeesBeesMob", true));
 
 					spawnProductiveBeesHoneycombVariants = subscriber.subscribe(builder
-		                    .comment(" \r\n-----------------------------------------------------\r\n\r\n"
-		                    		+" Spawn Productive Bees's various honeycomb variants in The Bumblezone\r\n"
-		                    		+" at all kinds of heights and height bands. Start exploring to find \r\n"
+		                    .comment(" \n-----------------------------------------------------\n\n"
+		                    		+" Spawn Productive Bees's various honeycomb variants in The Bumblezone\n"
+		                    		+" at all kinds of heights and height bands. Start exploring to find \n"
 		                    		+" where they spawn!"
-		                    		+" \r\n"
-		                    		+" NOTE: Will require a restart of the world to take effect. \r\n")
+		                    		+" \n"
+		                    		+" NOTE: Will require a restart of the world to take effect. \n")
 		                    .translation("the_bumblezone.config.modcompat.productivebees.spawnproductivebeeshoneycombvariants")
 		                    .define("spawnProductiveBeesHoneycombVariants", true));
 
 					allowHoneyTreatCompat = subscriber.subscribe(builder
-							.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-									+" Allow Honey Treat to be able to feed bees and Honeycomb Brood Blocks.\r\n")
+							.comment(" \n-----------------------------------------------------\n\n"
+									+" Allow Honey Treat to be able to feed bees and Honeycomb Brood Blocks.\n")
 							.translation("the_bumblezone.config.modcompat.productivebees.allowhoneytreatcompat")
 							.define("allowHoneyTreatCompat", true));
 
 
 	        		PBOreHoneycombSpawnRateBeeDungeon = subscriber.subscribe(builder
-	    		            .comment(" \r\n-----------------------------------------------------\r\n\r\n"
-	    		            		+" How much of Bee Dungeons is made of ore-based honeycombs.\r\n"
-	    		            		+" 0 is no or honeycombs, 1 is max ore honeycombs, and default is 0.3D\r\n")
+	    		            .comment(" \n-----------------------------------------------------\n\n"
+	    		            		+" How much of Bee Dungeons is made of ore-based honeycombs.\n"
+	    		            		+" 0 is no or honeycombs, 1 is max ore honeycombs, and default is 0.3D\n")
 	    		            .translation("the_bumblezone.config.productivebees.pborehoneycombspawnratebeedungeon")
 	    		            .defineInRange("PBOreHoneycombSpawnRateBeeDungeon", 0.3D, 0D, 1D));
 
 	        		PBGreatHoneycombRarityBeeDungeon = subscriber.subscribe(builder
-	    		            .comment(" \r\n-----------------------------------------------------\r\n\r\n"
-	    		            		+" How rare good ore-based Honeycombs (diamonds, ender, emerald, etc) are \r\n"
-	    		            		+" in Bee Dungeons. \r\n"
-	    		            		+" Higher numbers means more rare. Default rate is 3.\r\n")
+	    		            .comment(" \n-----------------------------------------------------\n\n"
+	    		            		+" How rare good ore-based Honeycombs (diamonds, ender, emerald, etc) are \n"
+	    		            		+" in Bee Dungeons. \n"
+	    		            		+" Higher numbers means more rare. Default rate is 3.\n")
 	    		            .translation("the_bumblezone.config.productivebees.pbgreathoneycombraritybeedungeon")
 	    		            .defineInRange("PBGreatHoneycombRarityBeeDungeon", 2, 1, 1001));
 
 	        		PBOreHoneycombSpawnRateSpiderBeeDungeon = subscriber.subscribe(builder
-	    		            .comment(" \r\n-----------------------------------------------------\r\n\r\n"
-	    		            		+" How much of Spider Infested Bee Dungeons is made of ore-based honeycombs.\r\n"
-	    		            		+" 0 is no or honeycombs, 1 is max ore honeycombs, and default is 0.1D\r\n")
+	    		            .comment(" \n-----------------------------------------------------\n\n"
+	    		            		+" How much of Spider Infested Bee Dungeons is made of ore-based honeycombs.\n"
+	    		            		+" 0 is no or honeycombs, 1 is max ore honeycombs, and default is 0.1D\n")
 	    		            .translation("the_bumblezone.config.productivebees.pborehoneycombspawnratespiderbeedungeon")
 	    		            .defineInRange("PBOreHoneycombSpawnRateSpiderBeeDungeon", 0.1D, 0D, 1D));
 
 	        		PBGreatHoneycombRaritySpiderBeeDungeon = subscriber.subscribe(builder
-	    		            .comment(" \r\n-----------------------------------------------------\r\n\r\n"
-	    		            		+" How rare good ore-based Honeycombs (diamonds, ender, emerald, etc) are \r\n"
-	    		            		+" in Spider Infested Bee Dungeons. \r\n"
-	    		            		+" Higher numbers means more rare. Default rate is 2.\r\n")
+	    		            .comment(" \n-----------------------------------------------------\n\n"
+	    		            		+" How rare good ore-based Honeycombs (diamonds, ender, emerald, etc) are \n"
+	    		            		+" in Spider Infested Bee Dungeons. \n"
+	    		            		+" Higher numbers means more rare. Default rate is 2.\n")
 	    		            .translation("the_bumblezone.config.productivebees.pbgreathoneycombrarityspiderbeedungeon")
 	    		            .defineInRange("PBGreatHoneycombRaritySpiderBeeDungeon", 2, 1, 1001));
 
 
 					PBBlacklistedBees = subscriber.subscribe(builder
-							.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-									+" Blacklist what Productive Bees bees should not spawn in Bumblezone. \r\n"
-									+" Separate each entry with a comma. Example: \"productivebees:iron,productivebees:coal\"\r\n"
-									+" \r\n"
-									+" Note: this is only for the entities. To blacklist blocks as well,\r\n"
-									+" use a datapack to add blacklisted blocks to this tag:\r\n"
-									+" data/the_bumblezone/tags/blocks/blacklisted_productive_bees_combs.json\r\n")
+							.comment(" \n-----------------------------------------------------\n\n"
+									+" Blacklist what Productive Bees bees should not spawn in Bumblezone. \n"
+									+" Separate each entry with a comma. Example: \"productivebees:iron,productivebees:coal\"\n"
+									+" \n"
+									+" Note: this is only for the entities. To blacklist blocks as well,\n"
+									+" use a datapack to add blacklisted blocks to this tag:\n"
+									+" data/the_bumblezone/tags/blocks/blacklisted_productive_bees_combs.json\n")
 							.translation("the_bumblezone.config.productivebees.pbblacklistedbees")
 							.define("PBBlacklistedBees", ""));
 
@@ -260,17 +260,17 @@ public class BzModCompatibilityConfigs
 	            builder.push("Potion of Bees Options");
 
 	            		allowPotionOfBeesCompat = subscriber.subscribe(builder
-		                    .comment(" \r\n-----------------------------------------------------\r\n\r\n"
-		                    		+" Allow Potion of Bees item to turn Empty Honeycomb Brood blocks \r\n"
-			                    	+" back into Honeycomb Brood Blocks with a larva in it. (affects Dispenser too)\r\n")
+		                    .comment(" \n-----------------------------------------------------\n\n"
+		                    		+" Allow Potion of Bees item to turn Empty Honeycomb Brood blocks \n"
+			                    	+" back into Honeycomb Brood Blocks with a larva in it. (affects Dispenser too)\n")
 		                    .translation("the_bumblezone.config.modcompat.potionofbees.allowpotionofbeescompat")
 		                    .define("allowPotionOfBeesCompat", true));
 
 	            		allowSplashPotionOfBeesCompat = subscriber.subscribe(builder
-			            .comment(" \r\n-----------------------------------------------------\r\n\r\n"
-			                    	+" Allow Splash Potion of Bees item to turn Empty Honeycomb Brood \r\n"
-				                +" blocks back into Honeycomb Brood Blocks with a larva in it when \r\n"
-				                +" the potion is thrown and splashed near the block. (affects Dispenser too)\r\n")
+			            .comment(" \n-----------------------------------------------------\n\n"
+			                    	+" Allow Splash Potion of Bees item to turn Empty Honeycomb Brood ",
+				               " blocks back into Honeycomb Brood Blocks with a larva in it when ",
+				               " the potion is thrown and splashed near the block. (affects Dispenser too)\n")
 			            .translation("the_bumblezone.config.modcompat.productivebees.allowsplashpotionofbeescompat")
 			            .define("allowSplashPotionOfBeesCompat", true));
 
@@ -280,63 +280,63 @@ public class BzModCompatibilityConfigs
 			builder.push("Buzzier Bees Options");
 
 				allowBottledBeesCompat = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" Allow Bottles Bees item to turn Empty Honeycomb Brood blocks \r\n"
-								+" back into Honeycomb Brood Blocks with a larva in it. (affects Dispenser too)\r\n")
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" Allow Bottles Bees item to turn Empty Honeycomb Brood blocks \n"
+								+" back into Honeycomb Brood Blocks with a larva in it. (affects Dispenser too)\n")
 						.translation("the_bumblezone.config.modcompat.buzzierbees.allowbottledbeescompat")
 						.define("allowBottledBeesCompat", true));
 
 				allowHoneyWandCompat = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" Allow Honey Wand to take honey from Filled Porous Honeycomb Block \r\n"
-								+" and put honey into Porous Honeycomb Block without angering bees.\r\n")
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" Allow Honey Wand to take honey from Filled Porous Honeycomb Block \n"
+								+" and put honey into Porous Honeycomb Block without angering bees.\n")
 						.translation("the_bumblezone.config.modcompat.buzzierbees.allowhoneywandcompat")
 						.define("allowHoneyWandCompat", true));
 
 				crystallizedHoneyWorldgen = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
+						.comment(" \n-----------------------------------------------------\n\n"
 								+" Place Buzzier Bees's Crystallized Honey Blocks on the /r/n"
-								+" surface of land around sea level and above.\r\n")
+								+" surface of land around sea level and above.\n")
 						.translation("the_bumblezone.config.modcompat.buzzierbees.crystallizedhoneyworldgen")
 						.define("crystallizedHoneyWorldgen", true));
 
 				allowRegularCandlesBeeDungeon = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
+						.comment(" \n-----------------------------------------------------\n\n"
 								+" Allow Bee Dungeons to have normal unscented candles./r/n")
 						.translation("the_bumblezone.config.modcompat.buzzierbees.allowregularcandlesbeedungeon")
 						.define("allowRegularCandlesBeeDungeon", true));
 
 				allowScentedCandlesBeeDungeon = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
+						.comment(" \n-----------------------------------------------------\n\n"
 								+" Allow Bee Dungeons to have scented candles that gives status effects./r/n")
 						.translation("the_bumblezone.config.modcompat.buzzierbees.allowscentedcandlesbeedungeon")
 						.define("allowScentedCandlesBeeDungeon", true));
 
 				allowScentedCandlesSpiderBeeDungeon = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
+						.comment(" \n-----------------------------------------------------\n\n"
 								+" Allow Spider Infested Bee Dungeons to have scented candles that gives status effects./r/n")
 						.translation("the_bumblezone.config.modcompat.buzzierbees.allowscentedcandlesspiderbeedungeon")
 						.define("allowScentedCandlesSpiderBeeDungeon", true));
 
 				powerfulCandlesRarityBeeDungeon = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" How rare are powerful candles in Bee Dungeons. \r\n"
-								+" Higher numbers means more rare.\r\n"
-								+" Default rate is 2.\r\n")
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" How rare are powerful candles in Bee Dungeons. \n"
+								+" Higher numbers means more rare.\n"
+								+" Default rate is 2.\n")
 						.translation("the_bumblezone.config.dungeons.powerfulcandlesraritybeedungeon")
 						.defineInRange("powerfulCandlesRarityBeeDungeon", 2, 0, 10));
 
 				powerfulCandlesRaritySpiderBeeDungeon = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" How rare are powerful candles in Spider Infested Bee Dungeons. \r\n"
-								+" Higher numbers means more rare.\r\n"
-								+" Default rate is 2.\r\n")
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" How rare are powerful candles in Spider Infested Bee Dungeons. \n"
+								+" Higher numbers means more rare.\n"
+								+" Default rate is 2.\n")
 						.translation("the_bumblezone.config.dungeons.powerfulcandlesrarityspiderbeedungeon")
 						.defineInRange("powerfulCandlesRaritySpiderBeeDungeon", 0, 0, 10));
 
 				allowBuzzierBeesTradeCompat = subscriber.subscribe(builder
-						.comment(" \r\n-----------------------------------------------------\r\n\r\n"
-								+" Add new trades to Buzzier Bees's Beekeeper Villager.\r\n")
+						.comment(" \n-----------------------------------------------------\n\n"
+								+" Add new trades to Buzzier Bees's Beekeeper Villager.\n")
 						.translation("the_bumblezone.config.modcompat.resourcefulbees.allowbuzzierbeestradecompat")
 						.define("allowBuzzierBeesTradeCompat", true));