+++ src/main/java/com/telepathicgrunt/the_bumblezone/mixin/LivingEntityAccessor.java
@@ -1,0 +1,11 @@
+package com.telepathicgrunt.the_bumblezone.mixin;
+
+import net.minecraft.entity.LivingEntity;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(LivingEntity.class)
+public interface LivingEntityAccessor {
+    @Accessor("jumping")
+    boolean isJumping();
+}